+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Notable changes to this project.
 
 ## [dev] - unreleased
 - include `reputationPayments` in `get_payments`
+- add datetime information (`insertedAt`) to `get_transactions`
 
 ## [1.5.4] - 2019-05-30
 - return new `reputation` as announced by numerai on 2019-05-29 in `get_rankings`